Moved Datatable object from base.py to its own file datatable.py
