+++ great_expectations/dataset/__init__.py
@@ -1,4 +1,5 @@
-from .base import Dataset, Datatable, DataFile
+from .base import Dataset, DataFile
+from .datatable import Datatable
 from .pandas_datatable import MetaPandasDatatable, PandasDatatable
 from .sqlalchemy_datatable import MetaSqlAlchemyDatatable, SqlAlchemyDatatable
 from .file_dataset import  MetaFileDataset, FileDataset
